python-algorithm/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
pymongo==4.3.3
redis==4.5.5
prometheus-client==0.16.0
pydantic==1.10.7
httpx==0.24.0
pytest==7.3.1
pytest-asyncio==0.21.0
mypy==1.3.0
black==23.3.0
isort==5.12.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0

